ENGINE v0.00000000000001
Just draws an image so far. But now some models, like the game loop and component system, are abstracted to an "engine" so it is on another layer and can be called by a developer. Also added a GameState functionality to the engine. Engine seems to be working so far.