Cross-map teleportation
A node-based system was added to allow teleportation BETWEEN tilemaps. Player would step on the node, it would unload the current tilemap and remove all its entities, then it would load the tilemap the node points to, and all the objects on that new tilemap, then place the player at the new node that the old node pointed to. The code for it is hard to follow, but is mostly contained in the onEnter and onExit for both the mainstate and loadworldstate modules.